requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
pyyaml
pytest
